fix(reconciler): throw errors for nil hook request or response
This commit throws error if either hook request or hook response
are nil. These objects are expected to be instantiated by metac.
Reconciliation will not work as expected if any one or both of
them are nil.

This commit also has additional scripts, artifacts required to
deploy or delete the entire cstorpoolauto setup.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>