fix(annotations): move to labels from annotations
This commit is a workaround due to a bug in metac. This modified
logic moves away from blockdevice annotations & uses labels in
its place.

NOTE:
Metac is not merging the annotations properly when reconciling a
resource to its desired state. This happens in GenericController
with updateAny set to true and the resource updateStratgey being
InPlace. In addition, it happens when some annotation is modified
in the targeted resource with above update tunables.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>