fix(raidgroupconfig): fix raidgroupconfig validation (#98)
* fix raidgroupconfig validation

There is one edge case in raidgroupconfig validation -
If device count is less than min device count then that is not a valid
raid group config. ie - For raid-z(2^n + 1) device count 1 is valid if
n=0 but that is not a valid raid group config. For raid-z2(2^n + 2)
device count 2 is valid if n=0 but that is not a valid raid group config.

This PR fixes this edge case in raidgroupconfig validation.

Signed-off-by: Shovan Maity <shovan.maity@mayadata.io>