fix(config): copy code from v1alpha1 package (#123)
Make separate config for v1 and v1alpha1 cspc. If we use one config and
crd present in the config is not present then it will stuck. If we use 2
configs then it will run on 2 separate goroutines. If cspc v1 crd is not
present then it will not block v1alpha1 cspc creation. If v1alpha1 crd
is not present then it will not block v1 cspc creation.

Signed-off-by: Shovan Maity <shovan.cse91@gmail.com>